Refactor import statements in CurrentWeather.tsx and weatherUtils.ts for consistency and readability
